create post issue

+++ app/Http/Controllers/Admin/CommentsController.php
@@ -19,15 +19,21 @@ class CommentsController extends Controller
     //评论列表
     public function index()
     {
-        $comments = $this->comment->getCommentsFeed();
-        return view('admin.comments.index',compact('comments'));
+        if(user()->hasRole('admin') || user()->hasRole('admin-two')){
+            $comments = $this->comment->getCommentsFeed();
+            return view('admin.comments.index',compact('comments'));
+        }
+        return redirect()->route('admin.index');
     }
 
     //删除评论
     public function destroy($id)
     {
-        $comment = $this->comment->byId($id);
-        $comment->delete();
-        return redirect()->route('admin.comments');
+        if(user()->hasRole('admin') || user()->hasRole('admin-two')){
+            $comment = $this->comment->byId($id);
+            $comment->delete();
+            return redirect()->route('admin.comments');
+        }
+        return redirect()->route('admin.index');
     }
 }
